requirements.txt
boto3==1.17.33
discord.py==1.6.0
python-dotenv==0.15.0
